backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.2
matplotlib==3.8.1
seaborn==0.13.0
pyarrow==14.0.0
great_expectations==0.17.19
openpyxl==3.1.2
xlrd==2.0.1
fastapi-jwt-auth==0.5.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.1
pydantic-settings==2.0.3
typing-extensions==4.8.0
fastapi-utils==0.2.1
sqlacodegen==3.0.0rc2
py2neo==2021.2.4
fastapi-pagination==0.12.10
strawberry-graphql==0.211.1
python-jose==3.3.0